基本完成收藏/喜欢+ 专辑/作品 功能

+++ Tumcca/Tumcca/src/com/pineapple/mobilecraft/tumcca/data/Works.java
@@ -22,7 +22,7 @@ public class Works {
     public String tags = "";
     public String title = "";
     public String description = "";
-    public int albumId = -1;
+    public long albumId = -1;
     public List<Integer> pictures = new ArrayList<Integer>();
 
     public static final Works NULL = new Works();
